pyproject.toml
[tool.poetry]
name = "zero_food_waste_jarczan"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
djangorestframework = "^3.15.0"
psycopg2 = "^2.9.9"
gunicorn = "^21.2.0"
pytest = "8.1.1"
pytest-django = "^4.8.0"
django-debug-toolbar = "^4.3.0"
invoke = "^2.2.0"
black = "^24.3.0"
flake8 = "6.1.0"
pyproject-flake8 = "^6.1.0"
pre-commit = "^3.6.2"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.dev-dependencies]
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '\.eggs, \.git, \.hg, \.mypy_cache, \.nox, \.tox, \.venv, _build, buck-out, build, dist'

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203"]
